requirements.txt
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-GraphQL==2.0.0
Flask-GraphQL-Auth==1.3.0
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
graphene==2.1.8
graphene-sqlalchemy==2.2.2
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.1
marshmallow==3.0.5
marshmallow-sqlalchemy==0.19.0
mimesis==3.3.0